Rename 'con' package to 'conflux' for Windows compatibility
The 'con' directory name is reserved in Windows (CON, PRN, AUX, NUL, etc. are
DOS device names), preventing users from cloning or using this project on
Windows systems. This renaming resolves the issue while preserving the set's
identity (CON = Conflux).

Updated all package declarations from com.wingedsheep.mtg.sets.definitions.con.*
to com.wingedsheep.mtg.sets.definitions.conflux.*
