+++ mtg-sets/src/main/kotlin/com/wingedsheep/mtg/sets/definitions/conflux/cards/UnsummonReprint.kt
@@ -1,4 +1,4 @@
-package com.wingedsheep.mtg.sets.definitions.con.cards
+package com.wingedsheep.mtg.sets.definitions.conflux.cards
 
 import com.wingedsheep.sdk.model.Printing
 import com.wingedsheep.sdk.model.Rarity
